requirements.txt
langchain
langgraph
pandas
langchain-experimental
langchain-google-genai
python-dotenv
streamlit